+++ src/test/run-pass/cast-region-to-uint.rs
@@ -8,9 +8,9 @@
 // option. This file may not be copied, modified, or distributed
 // except according to those terms.
 
-use std::ptr;
+use std::borrow;
 
 pub fn main() {
     let x = 3;
-    debug!("&x=%x", ptr::to_uint(&x));
+    debug!("&x=%x", borrow::to_uint(&x));
 }
